feat: point-core 공통 Exception 모델, 상수 추가

+++ point-core/src/test/java/com/github/kukim/point/core/domain/message/PointMessageTest.java
@@ -2,6 +2,7 @@ package com.github.kukim.point.core.domain.message;
 
 import com.github.kukim.point.core.domain.type.EventDetailType;
 import com.github.kukim.point.core.domain.type.EventType;
+import java.math.BigDecimal;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
 
@@ -10,7 +11,7 @@ class PointMessageTest {
 	@Test
 	void testConstructor() {
 		PointMessage pointMessage = PointMessage.create("0000001", EventType.SAVE,
-			EventDetailType.SAVE_EVENT, 100L, "로그인 적립",
+			EventDetailType.SAVE_EVENT, BigDecimal.valueOf(100L), "로그인 적립",
 			10L);
 
 		SoftAssertions.assertSoftly(softly -> {
